Advanced Relationships with Entity Framework

+++ Services/CharacterService/CharacterService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
 using dotnet_rpg.Data;
@@ -13,23 +14,30 @@ namespace dotnet_rpg.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
         
-        public CharacterService(IMapper mapper, DataContext context)
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+
+        private int getUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
             Character character = _mapper.Map<Character>(newCharacter);
-            
+            character.User = await _context.Users.FirstOrDefaultAsync(x=>x.Id == getUserId());
             _context.Characters.Add(character);
             await _context.SaveChangesAsync();
 
-            serviceResponse.Data = await _context.Characters.Select(x=> _mapper.Map<GetCharacterDto>(x)).ToListAsync();
+            serviceResponse.Data = await _context.Characters
+                                            .Where(x=>x.User.Id == getUserId())
+                                            .Select(x=> _mapper.Map<GetCharacterDto>(x)).ToListAsync();
+
             return serviceResponse;
         }
 
@@ -38,12 +46,21 @@ namespace dotnet_rpg.Services.CharacterService
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
             try {
-                Character character = _context.Characters.First(x=>x.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                Character character = await _context.Characters.FirstOrDefaultAsync(x=>x.Id == id && x.User.Id == getUserId());
 
-                serviceResponse.Data = _context.Characters.Select(x=> _mapper.Map<GetCharacterDto>(x)).ToList();
-            } 
+                if(character != null) {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+
+                    serviceResponse.Data = _context.Characters
+                    .Where(x=> x.User.Id == getUserId())
+                    .Select(x=> _mapper.Map<GetCharacterDto>(x)).ToList();
+                }
+                else {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character Not Found!";
+                }
+            }
             catch(Exception ex) {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
@@ -55,7 +72,11 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var response = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
+            var dbCharacters = await _context.Characters
+                    .Include(x => x.Weapon)
+                    .Include(x => x.Skills)
+                .Where(x => x.User.Id == getUserId())
+                .ToListAsync();
             response.Data = dbCharacters.Select(x => _mapper.Map<GetCharacterDto>(x)).ToList();
 
             return response;
@@ -64,7 +85,9 @@ namespace dotnet_rpg.Services.CharacterService
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(x=> x.Id == id);
+            var dbCharacter = await _context.Characters
+                    .Include(x => x.Weapon)
+                    .Include(x => x.Skills).FirstOrDefaultAsync(x=> x.Id == id && x.User.Id == getUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -74,18 +97,24 @@ namespace dotnet_rpg.Services.CharacterService
             ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
 
             try {
-                Character? character = await _context.Characters.FirstOrDefaultAsync(x=>x.Id == updateCharacterDto.Id);
+                Character? character = await _context.Characters
+                        .Include(x=>x.User)
+                        .FirstOrDefaultAsync(x=>x.Id == updateCharacterDto.Id);
 
-                // _mapper.Map(updateCharacterDto, character);
-                character.Name = updateCharacterDto.Name;
-                character.HitPoints = updateCharacterDto.HitPoints;
-                character.Defense = updateCharacterDto.Defense;
-                character.Strength = updateCharacterDto.Strength;
-                character.Intelligence = updateCharacterDto.Intelligence;
-                character.Class = updateCharacterDto.Class;
+                if(character.User.Id == getUserId()) {
+                    character.Name = updateCharacterDto.Name;
+                    character.HitPoints = updateCharacterDto.HitPoints;
+                    character.Defense = updateCharacterDto.Defense;
+                    character.Strength = updateCharacterDto.Strength;
+                    character.Intelligence = updateCharacterDto.Intelligence;
+                    character.Class = updateCharacterDto.Class;
 
-                await _context.SaveChangesAsync();
-                response.Data = _mapper.Map<GetCharacterDto>(character);
+                    await _context.SaveChangesAsync();
+                    response.Data = _mapper.Map<GetCharacterDto>(character);
+                } else {
+                    response.Success = false;
+                    response.Message = "Character Not Found!";
+                }
             } 
             catch(Exception ex) {
                 response.Success = false;
@@ -94,6 +123,48 @@ namespace dotnet_rpg.Services.CharacterService
 
             return response;
         }
-        
+
+        public async Task<ServiceResponse<GetCharacterDto>> AddCharacterSkill(AddCharacterSkillDto newCharacterSkill)
+        {
+            var response = new ServiceResponse<GetCharacterDto>();
+
+            try
+            {
+                var character = await _context.Characters
+                    .Include(x => x.Weapon)
+                    .Include(x => x.Skills)
+                    .FirstOrDefaultAsync(x=> x.Id == newCharacterSkill.CharacterId && x.User.Id == getUserId());
+
+                if(character == null)
+                {
+                    response.Success = false;
+                    response.Message = "Character Not Found!";
+
+                    return response;
+                }
+
+                var skill = await _context.Skills.FirstOrDefaultAsync(s => s.Id == newCharacterSkill.SkillId);
+
+                if(skill == null)
+                {
+                    response.Success = false;
+                    response.Message = "Skill Not Found!";
+
+                    return response;
+                }
+
+                character.Skills.Add(skill);
+                await _context.SaveChangesAsync();
+
+                response.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            catch(Exception ex)
+            {
+                response.Success = false;
+                response.Message = ex.Message;
+            }
+
+            return response;
+        }
   }
 }
